requirements.txt
connexion == 2.0.2
swagger-ui-bundle == 0.0.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
pyhumps >= 1.2.2
py2neo >= 2.5.0
healthcheck >= 1.3.3
iso8601 >= 0.1.12
pytz >= 2019.1
